refactor: Abstract the MemoryManager Logic. #2808
